Added another axios request to a new api endpoint to access state data, built out the navigation, added another table with the state data only, added an about page, added copyright footer, tweaked a lot of styling in CSS, downloaded some stuff, and now working on adding a weather api to the app
