+++ covid-19-app/package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "axios": "^0.21.1",
     "chart.js": "^2.9.3",
     "leaflet": "^1.7.1",
     "numeral": "^2.0.6",
@@ -14,7 +15,9 @@
     "react-chartjs-2": "^2.10.0",
     "react-dom": "^16.13.1",
     "react-leaflet": "^2.7.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
+    "reactstrap": "^8.9.0",
     "styled-components": "^5.2.0"
   },
   "scripts": {
